movie-service/requirements.txt
Django>=2.0,<3.0
psycopg2-binary>=2.7,<3.0
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
django-rest-swagger